✨ 핀 코멘트 목록 페이지에서 핀 이름 클릭시 핀 상세정보 모달이 팝업되는 기능 추가
